[aladdin_table.py] Don't use models for datapath ops
Data path ops have the bitwidth assumption built in. For example, fp32
is going to be 32-bit. It's weird to assume there might be a something
like fp20 in the world. Also, users might not remember to specify
datawidth under fp32/fp64 because it is strange to specify the bitwidth
for a single/double floating point op.